+++ doc/about/changes.md
@@ -7,6 +7,7 @@ in progress
 ### Bug fixes
 - Error reporting now shows the proper stack frames on Windows
 - Fix detection of `clang-cl` compiler
+- MinGW can now use libraries of the form `name.lib`
 
 ---